Move EllOrThetaConfig to generators
